requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1
requests==2.31.0
paystack==1.5.0
fpdf==1.7.2
python-magic==0.4.27
libmagic==1.0
python-magic-bin==0.4.14